+++ src/functions/numeric/leastCommonMultiple.js
@@ -17,7 +17,7 @@ class LimeFunctionLeastCommonMultiple extends LimeFunction {
 
     // Algorithms
     this.algorithms.set('r(arg{int[2]})', (step) => {
-      const l = step.right.value[0], r = step.right.value[1];
+      const l = step.right.value[0]; const r = step.right.value[1];
       step.rus(this.lime.direct([l, '*', r, '/', 'gcd', '(', l, ',', r, ')']));
     });
   }