+++ utils/banner.js
@@ -1,6 +1,10 @@
+// Require
+const Package = require('../package.json');
+
+/* ------------------------ division ------------------------ */
+
 // Banner function
 function banner() {
-  const Package = require('../package.json');
   return `Lime version ${Package.version}\n${Package.description}\n2021 ${Package.author}, ${Package.license} license\n${Package.homepage.slice(0, -7)}`;
 }