+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### v0.2.2 (dev)
+- Added expressions: `Matrix`
+- Added functions:
+  - Structural: `MatrixBracket`, `RowSplit`
 ### v0.2.1 (Aug 31, 2021)
 - Set up store system to prevent repeated calculation
 - Added expressions: `Complex`